WEA-6 Rename and move API services to new package
