+++ app/src/main/java/com/ayush/weatherapp/home/HomePresenterImpl.java
@@ -1,12 +1,12 @@
 package com.ayush.weatherapp.home;
 
 import android.support.annotation.NonNull;
-import com.ayush.weatherapp.home.pojo.CurrentForecast;
-import com.ayush.weatherapp.home.pojo.DailyForecast;
-import com.ayush.weatherapp.home.pojo.Forecast;
+import com.ayush.weatherapp.retrofit.weatherApi.pojo.CurrentForecast;
+import com.ayush.weatherapp.retrofit.weatherApi.pojo.DailyForecast;
+import com.ayush.weatherapp.retrofit.weatherApi.pojo.Forecast;
 import com.ayush.weatherapp.mvp.BaseContract;
-import com.ayush.weatherapp.utils.MapperUtils;
-import java.util.Calendar;
+import com.ayush.weatherapp.retrofit.weatherApi.WeatherAPIClient;
+import com.ayush.weatherapp.retrofit.weatherApi.WeatherAPIInterface;
 import java.util.List;
 import retrofit2.Call;
 import retrofit2.Callback;
@@ -28,14 +28,14 @@ public class HomePresenterImpl implements HomeContract.Presenter {
   @Override public void fetchWeatherDetails() {
     getView().showProgressDialog("Loading", false);
 
-    APIInterface apiInterface = APIClient.getClient().create(APIInterface.class);
+    WeatherAPIInterface weatherApiInterface = WeatherAPIClient.getClient().create(WeatherAPIInterface.class);
 
     //TODO get coordinates based on a location
     final double TEST_LATITUDE = 37.8267;
     final double TEST_LONGITUDE = -122.4233;
     String requestString = TEST_LATITUDE + "," + TEST_LONGITUDE;
 
-    Call<Forecast> forecastCall = apiInterface.getForecast(requestString);
+    Call<Forecast> forecastCall = weatherApiInterface.getForecast(requestString);
 
     forecastCall.enqueue(new Callback<Forecast>() {
       @Override
