+++ app/src/main/java/com/ayush/weatherapp/retrofit/weatherApi/WeatherAPIInterface.java
@@ -1,11 +1,11 @@
-package com.ayush.weatherapp.home;
+package com.ayush.weatherapp.retrofit.weatherApi;
 
-import com.ayush.weatherapp.home.pojo.Forecast;
+import com.ayush.weatherapp.retrofit.weatherApi.pojo.Forecast;
 import retrofit2.Call;
 import retrofit2.http.GET;
 import retrofit2.http.Path;
 
-public interface APIInterface {
+public interface WeatherAPIInterface {
 
   @GET("{coordinates}?exclude=minutely,alerts,flags")
   Call<Forecast> getForecast(@Path("coordinates") String coordinates);
