+++ app/src/main/java/com/ayush/weatherapp/utils/MathUtils.java
